renamed client_actions into CamelCase
