+++ pages/source/source_list.vue
@@ -30,7 +30,7 @@
 </template>
 
 <script>
-import ClientActions from '~/components/client_actions.vue'
+import ClientActions from '~/components/ClientActions.vue'
 import Source from '~/models/Source'
 
 export default {
